+++ src/Widgets/SettingsDialog.h
@@ -82,6 +82,7 @@ class WIDGETS_EXPORT SettingsDialog : public QDialog, Ui::SettingsDialog
         QCheckBox* checkBoxPreviewAudioMeters;
         QCheckBox* checkBoxPreviewTemplates;
         QCheckBox* checkBoxPreviewLegacyMode;
+        QCheckBox* checkBoxOgrafEnabled;
         QLabel* labelAutoSaveMinutes;
         QSpinBox* spinBoxAutoSaveMinutes;
         QCheckBox* checkBoxShowLastAction;